% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{buildRadialSetsPlot}
\alias{buildRadialSetsPlot}
\title{Build radial sets plot}
\usage{
buildRadialSetsPlot(setSizes, setSizesByDegree, setIntersections,
  linkThickness = "percent", focusSet = "none",
  sectorLabelFontSize = 1, axisLabelFontSize = 1,
  maxLinkThickness = 15, countScale = 1, facing = "clockwise",
  axisLabels = FALSE, legendTitle = NULL, showLegend = FALSE,
  sectorLineWidth = 2, sectorColor = "white", majorTick = FALSE)
}
\description{
Build radial sets plot
}
