% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{createRadialsetsTooltip}
\alias{createRadialsetsTooltip}
\title{Tooltip for radial sets plot}
\usage{
createRadialsetsTooltip(setSizes, setSizesByDegree, setIntersections,
  pointerLoc, focusSets = "none", linkThickness = "prop")
}
\arguments{
\item{setSizes}{Data frame of set sizes from \code{\link{getSetSizes}}}

\item{setSizesByDegree}{Data frame of set sizes by degree from
\code{\link{getSetSizesByDegree}}}

\item{setIntersections}{Data frame of set intersection sizes from
\code{\link{getSetIntersections}}}

\item{pointerLoc}{A list from \code{\link{getPointerLoc}}}

\item{focusSets}{Character vector specifying name(s) of set(s) to focus on.
If single set is specified, only links originating from set are shown. If
multiple sets are specified, only links between sets are shown.}

\item{linkThickness}{Character vector specifying name of variable from
\code{setIntersections} data frame to map to link thickness}
}
\value{
A tooltip created as a panel using \code{\link[shiny]{wellPanel}}
}
\description{
Build tooltip for radial sets network plot
}
