% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{getPointerLoc}
\alias{getPointerLoc}
\title{Get Radial Sets pointer location}
\usage{
getPointerLoc(metadata, pointer, plotDomain = list(left = -1.08, right =
  1.08, bottom = -1.08, top = 1.08), transCoord = FALSE)
}
\arguments{
\item{plotData}{A data frame with data used in plot}

\item{hover}{A named list created by \code{\link[shiny]{plotOutput}}}
}
\value{
A tooltip created as a panel using \code{\link[shiny]{wellPanel}}
}
\description{
Build tooltip for radial sets network plot
}
