% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getRadialSetsData}
\alias{getRadialSetsData}
\title{Prepare data for Radial Sets plot}
\usage{
getRadialSetsData(setSizes, setSizesByDegree, setIntersections,
  setOrder = NULL, linkThickness = "prop",
  linkColor = "prop.RelError", linkColorPal = "RdBu",
  focusSets = "none", countScale = 1, colorScaleLim = c(-1, 1),
  reverseLinkPal = FALSE)
}
\arguments{
\item{setSizes}{Data frame of set sizes from \code{\link{getSetSizes}}}

\item{setSizesByDegree}{Data frame of set sizes by degree from
\code{\link{getSetSizesByDegree}}}

\item{setIntersections}{Data frame of set intersection sizes from
\code{\link{getSetIntersections}}}

\item{linkThickness}{Character vector specifying name of variable from
\code{setIntersections} data frame to map to link thickness}

\item{linkColor}{Character vector specifying name of variable from
\code{setIntersections} data frame to map to link color}

\item{linkColorPal}{Name of color palette to use for link colors. Must match
palette from \code{\link[RColorBrewer]{brewer.pal}}}

\item{focusSets}{Character vector specifying name(s) of set(s) to focus on.
If single set is specified, only links originating from set are shown. If
multiple sets are specified, only links between sets are shown.}

\item{countScale}{An optional input to apply multiplicative scaling to set
sizes}

\item{colorScaleLim}{An optional input to specify scale limits on link colors.
Values outside limits are mapped to ends of color scale.}
}
\description{
Create a list with data needed by \code{\link{buildRadialSetsPlot}}.
}
