% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{getRadialSetsMetadata}
\alias{getRadialSetsMetadata}
\title{Get Radial Sets metadata}
\usage{
getRadialSetsMetadata(radialSetsData, bezierW = 1, bezierHRatio = 0.5)
}
\arguments{
\item{plotData}{A data frame with data used in plot}

\item{hover}{A named list created by \code{\link[shiny]{plotOutput}}}
}
\value{
A tooltip created as a panel using \code{\link[shiny]{wellPanel}}
}
\description{
Build tooltip for radial sets network plot
}
