% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getSetSizes}
\alias{getSetSizes}
\title{Count total number of items in each set}
\usage{
getSetSizes(df, setNames)
}
\arguments{
\item{df}{A data frame indicating set membership}

\item{setNames}{A character vector of set names}
}
\value{
A data frame with variables \code{set} and \code{N} indicating the
number of items in each set
}
\description{
Count total number of items in each set
}
\details{
The input data frame should contain a row for each item and a binary variable
for each set indicating the membership of each item. The \code{setNames}
input should correspond to the binary indicator columns in the data frame.
}
\examples{
# Define set names
data("movieSets")
setNames <- movieSets \%>\%
  select(Action:Western) \%>\%
  colnames()

# Calculate set sizes
getSetSizes(movieSets, setNames)

}
