% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getSetSizesByDegree}
\alias{getSetSizesByDegree}
\title{Count number of items in each set by degree}
\usage{
getSetSizesByDegree(df, setNames, idName, maxDegree = 4)
}
\arguments{
\item{df}{A data frame indicating set membership}

\item{setNames}{A character vector of set names}

\item{idName}{A string specifying name of ID variable for each item}

\item{maxDegree}{A numeric input indicating upper limit on degree}
}
\value{
A data frame with variables:
\itemize{
\item \code{set} indicating set
\item \code{degree} indicating degree of overlap (maximum of \code{maxDegree})
\item \code{degreeLabel} a factor labeling the degree variable
\item \code{N} number of items
}
}
\description{
Count  number of items in each set by degree of overlap. The degree of
overlap is the number of sets of which an item is a member.
}
\details{
The input data frame should contain a row for each item and a binary variable
for each set indicating the membership of each item. The \code{setNames}
input should correspond to the binary indicator columns in the data frame.
}
\examples{
# Define set names
data("movieSets")
setNames <- movieSets  \%>\%
  select(Action:Western) \%>\%
  colnames()

# Calculate set sizes
getSetSizesByDegree(movieSets , setNames, "movieId")

# Calculate set sizes with max degree 3
getSetSizesByDegree(movieSets , setNames, "movieId", maxDegree = 3)

}
