% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groceries}
\alias{groceries}
\title{Point-of-sale transaction data from grocery outlet}
\format{A data frame with 9835 rows and 170 variables:
\describe{
  \item{frankfurter:bags}{logical columns indicating set membership}
  \item{id}{a transaction id based on row number}
}}
\source{
\url{https://cran.r-project.org/package=arules}
}
\usage{
groceries
}
\description{
This dataset was obtained from the arules package
(\url{https://cran.r-project.org/package=arules}) and converted
to a data frame. Please see \code{\link[arules]{Groceries}} in the arules
package and cite accordingly.
}
\references{
Michael Hahsler, Kurt Hornik, and Thomas Reutterer (2006) Implications of
 probabilistic data modeling for mining association rules.
 In M. Spiliopoulou, R. Kruse, C. Borgelt, A. Nuernberger, and W. Gaul,
  editors, From Data and Information Analysis to Knowledge Engineering,
  Studies in Classification, Data Analysis, and Knowledge Organization,
  pages 598–605. Springer-Verlag.
}
\keyword{datasets}
