% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movieSets}
\alias{movieSets}
\title{Genre classifications of 27,000 movies}
\format{A data frame with 27278 rows and 27 variables:
\itemize{
  \item{movieId}{MovieLens id (https://movielens.org/movies/\code{movieId})}
  \item{title}{movie title}
  \item{year}{year of release}
  \item{imdbId}{IMDB id (http://www.imdb.com/title/tt\code{imdbId})}
  \item{tmdbId}{The Movie DB id (https://www.themoviedb.org/movie/\code{tmdbId})}
  \item{avgRating}{average movie rating}
  \item{nRating}{number of ratings}
  \item{genres}{movie genres as string separated by "|"}
  \item{Action:Western}{binary columns indicating set membership}
}}
\source{
\url{https://grouplens.org/datasets/movielens/20m/}
}
\usage{
movieSets
}
\description{
A dataset containing the genres and other attributes of 27,000
movies.
}
\keyword{datasets}
