% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareRadSetsData.R
\name{prepareRadSetsData}
\alias{prepareRadSetsData}
\title{Prepare data for radial sets plot}
\usage{
prepareRadSetsData(sharedCustomers, degreeCount,
  linkThickness = "percent", focusGroup = "none",
  countScale = 1/1000)
}
\description{
Prepare data for radial sets plot
}
