% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{buildRadialSetsPlot}
\alias{buildRadialSetsPlot}
\title{Build radial sets plot}
\usage{
buildRadialSetsPlot(setSizes, setSizesByDegree, setIntersections,
  setOrder = NULL, linkThickness = "prop",
  linkColor = "prop.relError", linkColorPal = "RdBu",
  reverseLinkPal = FALSE, focusSets = "none", countScale = 1,
  colorScaleLim = c(-1, 1), colorScaleMapFun = "squish",
  edgeScaleLim = c(-Inf, Inf), edgeScaleMapFun = "censor",
  edgeWidthRange = c(1, 8), dropSets = FALSE,
  sectorLabelFontSize = 1, axisLabelFontSize = 1,
  facing = "clockwise", axisLabels = FALSE, legendTitle = NULL,
  showLegend = FALSE, sectorLineWidth = 1, majorTick = FALSE,
  bezierW = 1, bezierHRatio = 0.75, sectorColor = "white",
  barColor = "darkgrey")
}
\arguments{
\item{setSizes}{Data frame of set sizes from \code{\link{getSetSizes}}}

\item{setSizesByDegree}{Data frame of set sizes by degree from
\code{\link{getSetSizesByDegree}}}

\item{setIntersections}{Data frame of set intersection sizes from
\code{\link{getSetIntersections}}}

\item{setOrder}{An optional input to specify order of sets (counterclockwise).
Partial ordering is performed using \code{\link[forcats]{fct_relevel}}}

\item{linkThickness}{Character vector specifying name of variable from
\code{setIntersections} data frame to map to link thickness}

\item{linkColor}{Character vector specifying name of variable from
\code{setIntersections} data frame to map to link color}

\item{linkColorPal}{Name of color palette to use for link colors. Must match
palette from \code{\link[RColorBrewer]{brewer.pal}}}

\item{reverseLinkPal}{A logical input specifying if color scale should be reversed}

\item{focusSets}{Character vector specifying name(s) of set(s) to focus on.
If single set is specified, only links originating from set are shown. If
multiple sets are specified, only links between sets are shown.}

\item{countScale}{An optional input to apply multiplicative scaling to set
sizes}

\item{colorScaleLim}{A numeric vector specifying upper and lower scale limits
on link colors. Scale limits are enforced using \code{\link[scales]{censor}} or
\code{\link[scales]{squish}}}

\item{colorScaleMapFun}{A character string specifying whether to use
\code{\link[scales]{censor}} or \code{\link[scales]{squish}} to enforce color
scale limits}

\item{edgeScaleLim}{A numeric vector specifying upper and lower scale limits
on edge thickness. Scale limits are enforced using \code{\link[scales]{censor}} or
\code{\link[scales]{squish}}}

\item{edgeScaleMapFun}{A character string specifying whether to use
\code{\link[scales]{censor}} or \code{\link[scales]{squish}} to enforce edge
thickness limits}

\item{edgeWidthRange}{A numeric vector specifying minimum and maximum line
thickness of links}

\item{dropSets}{A logical indicating if sets without links should be removed}

\item{sectorLabelFontSize}{font size of sector labels}

\item{axisLabelFontSize}{font size of axis labels}

\item{facing}{direction of sector labels}

\item{axisLabels}{logical indicating if axis labels are show}

\item{legendTitle}{logical indicating if legend should have title}

\item{showLegend}{logical indicating if legend should be shown}

\item{sectorLineWidth}{line width of sector border}

\item{majorTick}{logical indicating if tick marks are shown}

\item{bezierW}{w parameter for link curves}

\item{bezierHRatio}{height ratio for link curves}

\item{sectorColor}{color of sectors}

\item{barColor}{color of bars}
}
\description{
Build radial sets plot
}
