% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{buildRadialSetsPlot}
\alias{buildRadialSetsPlot}
\title{Build radial sets plot}
\usage{
buildRadialSetsPlot(setSizes, setSizesByDegree, setIntersections,
  setOrder = NULL, linkThickness = "prop",
  linkColor = "prop.relError", linkColorPal = "RdBu",
  focusSets = "none", sectorLabelFontSize = 1, axisLabelFontSize = 1,
  maxLinkThickness = 15, countScale = 1, facing = "clockwise",
  axisLabels = FALSE, legendTitle = NULL, showLegend = FALSE,
  sectorLineWidth = 1, sectorColor = "white", majorTick = FALSE,
  bezierW = 1, bezierHRatio = 0.75, colorScaleLim = c(-1, 1),
  barColor = "darkgrey", reverseLinkPal = FALSE)
}
\description{
Build radial sets plot
}
