% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{getPointerLoc}
\alias{getPointerLoc}
\title{Get Radial Sets pointer location}
\usage{
getPointerLoc(metadata, pointer, plotDomain = list(left = -1.08, right =
  1.08, bottom = -1.08, top = 1.08), transCoord = FALSE)
}
\arguments{
\item{metadata}{A list of data frames from
\code{\link{getRadialSetsMetadata}}}

\item{pointer}{A named list created by \code{\link[shiny]{plotOutput}}}

\item{plotDomain}{A list of dimensions for converting pointer location on
image to plot coordinates}

\item{transCoord}{A logical indicating if coodinates from \code{pointer} should be
transformed based on \code{plotDomain}}
}
\description{
Build tooltip for radial sets network plot
}
