% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getRadialSetsData}
\alias{getRadialSetsData}
\title{Prepare data for Radial Sets plot}
\usage{
getRadialSetsData(setSizes, setSizesByDegree, setIntersections,
  setOrder = NULL, linkThickness = "prop",
  linkColor = "prop.relError", linkColorPal = "RdBu",
  focusSets = "none", countScale = 1, colorScaleLim = c(-1, 1),
  edgeWidthLim = NULL, reverseLinkPal = FALSE, maxPlotWidth = 15,
  minPlotWidth = 1)
}
\arguments{
\item{setSizes}{Data frame of set sizes from \code{\link{getSetSizes}}}

\item{setSizesByDegree}{Data frame of set sizes by degree from
\code{\link{getSetSizesByDegree}}}

\item{setIntersections}{Data frame of set intersection sizes from
\code{\link{getSetIntersections}}}

\item{setOrder}{An optional input to specify order of sets (counterclockwise).
Partial ordering is performed using \code{\link[forcats]{fct_relevel}}}

\item{linkThickness}{Character vector specifying name of variable from
\code{setIntersections} data frame to map to link thickness}

\item{linkColor}{Character vector specifying name of variable from
\code{setIntersections} data frame to map to link color}

\item{linkColorPal}{Name of color palette to use for link colors. Must match
palette from \code{\link[RColorBrewer]{brewer.pal}}}

\item{focusSets}{Character vector specifying name(s) of set(s) to focus on.
If single set is specified, only links originating from set are shown. If
multiple sets are specified, only links between sets are shown.}

\item{countScale}{An optional input to apply multiplicative scaling to set
sizes}

\item{colorScaleLim}{An optional input to specify scale limits on link colors.
Values outside limits are mapped to ends of color scale.}

\item{reverseLinkPal}{A logical input specifying if color scale should be reveresed}

\item{maxPlotWidth}{A numeric input specifying maximum line thickness of links}

\item{minPlotWidth}{A numeric input specifying minimum line thickness of links}

\item{edgeScaleLim}{An optional input to specify scale limits on link thickness.
Values outside limits are mapped to ends of thickness scale.}
}
\value{
A list with elements:
\itemize{
 \item \code{edgeWidth} matrix of link thickness variable (original scale)
 \item \code{edgeWidthMap} matrix of link thickness mapped to line thickness
 \item \code{edgeColor} matrix of link color variable (original scale)
 \item \code{edgeWidthMap} matrix of link color mapped to colors
 \item \code{degreeMat} matrix of bar sizes
 \item \code{sets} character vector of set names
 \item \code{nSets} number of sets
 \item \code{maxDegree} max degree of overlaps in bar plot
 \item \code{setSizesVec} vector of set sizes
 \item \code{maxWidth} maximum width of lines for edges
}
}
\description{
Create a list with data needed by \code{\link{buildRadialSetsPlot}}.
}
