% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getRadialSetsData}
\alias{getRadialSetsData}
\title{Prepare data for Radial Sets plot}
\usage{
getRadialSetsData(setSizes, setSizesByDegree, setIntersections,
  linkThickness = "percent", focusSet = "none", countScale = 1)
}
\arguments{
\item{setSizes}{Data frame of set sizes from \code{\link{getSetSizes}}}

\item{setSizesByDegree}{Data frame of set sizes by degree from
\code{\link{getSetSizesByDegree}}}

\item{setIntersections}{Data frame of set intersection sizes from
\code{\link{getSetIntersections}}}

\item{linkThickness}{Character vector specifying if edge matrix should be
a count or a proportion. Valid values are "percent" or "count"}

\item{focusSet}{Character vector specifying a name of a set to focus on.
Edges not connected to focus set are removed.}

\item{countScale}{An optional input to apply multiplicative scaling to set
sizes}
}
\description{
Create a list with data needed by \code{\link{buildRadialSetsPlot}}.
}
