% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{getRadialSetsMetadata}
\alias{getRadialSetsMetadata}
\title{Get Radial Sets metadata}
\usage{
getRadialSetsMetadata(radialSetsData, bezierW = 1, bezierHRatio = 0.5)
}
\arguments{
\item{radialSetsData}{Data frame from \code{\link{getRadialSetsData}}}

\item{bezierW}{w parameter for link curves}

\item{bezierHRatio}{height ratio for link curves}
}
\description{
Build tooltip for radial sets network plot
}
