% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{getSetIntersections}
\alias{getSetIntersections}
\title{Calculate sizes of intersections between sets}
\usage{
getSetIntersections(df, setNames, idName)
}
\arguments{
\item{df}{A data frame indicating set membership}

\item{setNames}{A character vector of set names}

\item{idName}{A string specifying name of ID variable for each item}
}
\value{
A data frame with variables:
\itemize{
 \item \code{set1} and \code{set2} pair of sets
 \item \code{Ninter} and \code{Nunion} number of elements in
 the intersection and union of the pair of sets
 \item \code{N1} and \code{N2} number of items in each set
 \item \code{prop1} and \code{prop2} proportion of items in
 each set that are also members of the other set
  (\code{prop1 = Ninter / N1})
 \item \code{prop} proportion of items from either set that
 are members of both sets (\code{prop = Ninter / Nunion})
 \item \code{.pred} predictions assuming marginal independence of sets
 \item \code{.error} difference between predictions and observations
 \item \code{.relError} error relative to observations
}
}
\description{
Calculate sizes of unions and intersections between each pair of sets
}
\details{
The input data frame should contain a row for each item and a binary variable
for each set indicating the membership of each item. The \code{setNames}
input should correspond to the binary indicator columns in the data frame.
}
\examples{
# Define set names
data(movies)
setNames <- movies \%>\%
  select(Action:Western) \%>\%
  colnames()

# Calculate set sizes
getSetIntersections(movies, setNames, "Name")

}
