% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_link_order.R
\name{optimizeLinkOrder}
\alias{optimizeLinkOrder}
\title{Optimize link order}
\usage{
optimizeLinkOrder(setSizes, setSizesByDegree, setIntersections,
  linkThickness = "prop")
}
\description{
Optimize link order to minimize ink
}
